requirements.txt
nemoguardrails>=0.10.0
openai>=1.0.0
chainlit>=2.0.0
deepteam>=1.0.6
python-dotenv>=1.0.0
deepeval>=3.0.0